Drop legacy log parsing code
