+++ setup/steps/li_ensure_supported_log_line_prefix.go
@@ -3,9 +3,11 @@ package steps
 import (
 	"errors"
 	"fmt"
+	"strings"
 
 	survey "github.com/AlecAivazis/survey/v2"
 	"github.com/lib/pq"
+	"github.com/pganalyze/collector/logs"
 	s "github.com/pganalyze/collector/setup/state"
 	"github.com/pganalyze/collector/setup/util"
 )
@@ -21,7 +23,12 @@ var EnsureSupportedLogLinePrefix = &s.Step{
 		}
 
 		currValue := row.GetString(0)
-		needsUpdate := !util.Includes(s.SupportedLogLinePrefixes, currValue) ||
+		hasDb := strings.Contains(currValue, "%d")
+		hasUser := strings.Contains(currValue, "%u")
+		hasTs := strings.Contains(currValue, "%m") || strings.Contains(currValue, "%n") || strings.Contains(currValue, "%t")
+		supported := hasDb && hasUser && hasTs
+
+		needsUpdate := !supported ||
 			(state.Inputs.Scripted &&
 				state.Inputs.GUCS.LogLinePrefix.Valid &&
 				currValue != state.Inputs.GUCS.LogLinePrefix.String)
@@ -35,36 +42,25 @@ var EnsureSupportedLogLinePrefix = &s.Step{
 				return errors.New("log_line_prefix not provided and current setting is not supported")
 			}
 			selectedPrefix = state.Inputs.GUCS.LogLinePrefix.String
-			if !util.Includes(s.SupportedLogLinePrefixes, selectedPrefix) {
-				return fmt.Errorf("log_line_prefix provided as unsupported value '%s'", selectedPrefix)
-			}
 		} else {
 			row, err := state.QueryRunner.QueryRow(`SELECT setting FROM pg_settings WHERE name = 'log_line_prefix'`)
 			if err != nil {
 				return err
 			}
 			oldVal := row.GetString(0)
-			var opts []string
-			for i, llp := range s.SupportedLogLinePrefixes {
-				// N.B.: we quote the options because many prefixes end in whitespace; we need to make that clear
-				var opt string
-				if i == 0 {
-					opt = fmt.Sprintf("'%s' (recommended)", llp)
-				} else {
-					opt = fmt.Sprintf("'%s'", llp)
-				}
-				opts = append(opts, opt)
-			}
-			var prefixIdx int
-			err = survey.AskOne(&survey.Select{
-				Message: fmt.Sprintf("Setting 'log_line_prefix' is set to unsupported value '%s'; set to (will be saved to Postgres):", oldVal),
-				Help:    "Check format specifier reference in Postgres documentation: https://www.postgresql.org/docs/current/runtime-config-logging.html#GUC-LOG-LINE-PREFIX",
-				Options: opts,
-			}, &prefixIdx)
+			err = survey.AskOne(&survey.Input{
+				Message: fmt.Sprintf("Setting 'log_line_prefix' (%s) is missing user (%%u), database (%%d), or timestamp (%%n, %%m, or %%t); set to (will be saved to Postgres):", oldVal),
+				Suggest: func(toComplete string) []string {
+					if toComplete == "" {
+						return []string{logs.LogPrefixRecommended}
+					}
+					return []string{}
+				},
+				Help: "Check format specifier reference in Postgres documentation: https://www.postgresql.org/docs/current/runtime-config-logging.html#GUC-LOG-LINE-PREFIX",
+			}, &selectedPrefix)
 			if err != nil {
 				return err
 			}
-			selectedPrefix = s.SupportedLogLinePrefixes[prefixIdx]
 		}
 		return util.ApplyConfigSetting("log_line_prefix", pq.QuoteLiteral(selectedPrefix), state.QueryRunner)
 	},
